requirements.txt
# For all scripts
openai
python-dotenv

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

mcp[cli]